products-service/requirements.txt
boto3
python-decouple
